requirements.txt
pyqt6
pyserial
inputs
opencv-python
pyqtgraph
configparser
numpy
pygame
pyyaml
sqlite3
matplotlib
Pillow
asyncio
websockets
flask
